s3 logic in base/message
