+++ lib/queuel/base/message.rb
@@ -76,6 +76,67 @@ module Queuel
       def decode_body?
         !decoder.nil? && decode? && raw_body.is_a?(String)
       end
+
+      def max_bytesize
+        options[:max_bytesize] || 64 * 1024
+      end
+
+      def push_message_body
+        if encoded_body.bytesize > max_bytesize
+          key = generate_key
+          s3_transaction(:write, key, encoded_body)
+          self.body = { 'queuel_s3_object' => key }
+        end
+        encoded_body
+      end
+
+      def pull_message_body
+        raise NotImplementedError, "must define method #pull_message_body"
+      end
+
+      def s3
+        @s3 ||= ::AWS::S3.new(
+                  :access_key_id => options[:s3_access_key_id],
+                  :secret_access_key => options[:s3_secret_access_key] )
+      end
+
+      # @method - write or read
+      # @args - key and message if writing
+      def s3_transaction(method, *args)
+        bucket_name = options[:s3_bucket_name]
+        raise NoBucketNameSupplied if bucket_name.nil?
+        my_bucket = s3.buckets[bucket_name]
+        if my_bucket.exists?
+          begin
+            send("s3_#{method}", my_bucket, *args)
+          rescue AWS::S3::Errors::AccessDenied => e
+            raise InsufficientPermissions, "Unable to read from bucket: #{e.message}"
+          end
+        else
+          raise BucketDoesNotExistError, "Bucket has either expired or does not exist"
+        end
+      end
+
+      def s3_read(bucket, *args)
+        bucket.objects[args[0]].read
+      end
+
+      def s3_write(bucket, *args)
+        bucket.objects[args[0]].write(args[1])
+      end
+
+      def generate_key
+        key = [
+          (Time.now.to_f * 10000).to_i,
+          SecureRandom.urlsafe_base64,
+          Thread.current.object_id
+        ].join('-')
+        key
+      end
+
+      class NoBucketNameSupplied < Exception; end
+      class InsufficientPermissions < StandardError; end
+      class BucketDoesNotExistError < StandardError; end
     end
   end
 end
